fix: change actionTypes to be enum, rename AuthProvider.tsx to userReducer.tsx
